Fix bg tremor when scroll to bottom on phone
When we scroll down the page on phone, the taskbar of the browser will hide automatically, and the height and weight of the device will be altered. Therefore, the bg height and weight will also be changed when we are using % on height and weight. The solution is change % to vh and vw.